hotfix: handle numeric GPL-ID column
A bug was found while adding GPL annotation data to an ExpressionSet.

Suppose a GPL annotation dataset contains a numeric ID column and we
want to pull out all rows of an ExpressionSet that contain a probe which
is present in the GPL annotation dataset. The probe IDs in the rownames
of the ExpressionSet will be strings, and we have to be careful that
comparison against the numeric GPL column doesn't convert the probe IDs
to numbers (because probe "987654" won't necessarily be in the 987654th
row of the ExpressionSet, if that row even exists)