get objects from string in MicroarrayPreprocessConfig()
In a MicroarrayPreprocessConfig object, the keep_sample_fn,
keep_probe_fn and entrezgene_db objects should be functions (in the
first two cases) or OrgDb objects (in the latter).

Here we allow the user to specify these objects by their name in the
form `some_pkg::some_object` or `some_object` (when the objects are
in the calling env).