add function 'design_builder'
This builder-function returns a function that converts an ExpressionSet
to a design data-frame.

A previously existing function `gld_fnBuilder_exptDesign` returned
functions that could act on `eset_limma_dataset` or `ExpressionSet`.
That function has been soft deprecated (it was poorly named and
complicated to use; it now throws a warning when the returned function
is used) and has been reimplemented to call `design_builder`.

Tests:
- the returned function acts on ExpressionSets
- the returned function fails if it creates a design matrix where the
row number doesn't match the number of samples in the ExpressionSet.